added utils.h and utils.cpp
