--- user.cpp
@@ -1,9 +1,0 @@
-#include <iostream>
-#include <fstream>
-#include <string>
-#include "user.h"
-using namespace std;
-
-User::User(): email(""), name(""), bio(""), friends(nullptr), likedPages(nullptr) {}
-
-User::User(string e): email(e) {}
